# API/OpenAPI/openapi.yaml
openapi: 3.0.0
info:
  title: Soupify™
  description: "This is a backend server API for storing recipes and meals we coocked.\
    \ \n*RestfulAPI* for the re-branding project of Engineering-Students-Recipes Company\
    \ to their new company, **Soupify™**. This project was creating as a group project\
    \ for Web-Development-Environments course *@Ben-Gurion University*. For more details\
    \ on the process/design please read the linked *README.md* file.\n  \n**Authorization**:\n\
    A logged-in user must update/delete his resources **ONLY**! Otherwise, he will\
    \ receive 403 Unauthorized Error.\n\n**Base URL:** `{your_custom_url}/api`\n\n\
    ---\n**Validity**:\nAll parameters sent to any API endpoint are validated according\
    \ to the following criteria:\n* Required vs Optional\n* Numeric vs Alpha vs Alphanumeric\n\
    * Unique vs Multiple\n* Min/max length of strings\n* Min/max values of numbers\n\
    * Date formats\n\n---\n**Request Headers**:\nAll request headers must have the\
    \ following fields:\n```\n  Authorization: Bearer <user's auth JSON Web Token>\n\
    ```\nYou can use this header token for admin access:\n\n---\n**Pagination**:\n\
    Pages and Limits are query parameters which are designed for paginate freely.\
    \ Pages and Limits must be **NUMERIC**.\n* page is the page number jump to.\n\
    * Limit is the number of items per page.\n* offest calculate as (page - 1) x limit\n\
...
    \ object *\n      }\n  ]\n}\n```\n---"
  contact:
...
servers:
...
paths:
...
  /recipes/:
    get:
      tags:
        - recipes
      summary: List all recipes
      operationId: listRecipes
      parameters:
        - name: page
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: integer
            format: int64
        - name: limit
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecipeWithId'
    post:
      tags:
        - recipes
      summary: Create a new recipe.
      operationId: createRecipe
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Recipe'
        required: true
      responses:
        "201":
          description: recipe was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeWithId'
        "400":
          description: Properties not set.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "403":
          description: Forbidden. Only admin can create
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "500":
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - bearerAuth: []
  /recipes/{id}:
    get:
      tags:
        - recipes
      summary: Get a recipe by {id}.
      description: Retrive a recipe from database or from spooncaular API with given
        id according to local parameter. local=true => DB
      operationId: getDishById
      parameters:
        - name: id
          in: path
          description: recipe id.
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            example: 151
        - name: local
          in: query
          description: recipe id.
          required: false
          style: form
          explode: true
          schema:
            type: boolean
            example: true
      responses:
        "200":
          description: OK
          content:
            application-json:
              schema:
                $ref: '#/components/schemas/RecipeWithId'
        "401":
          description: Unauthorized.
        "404":
          description: recipe not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "500":
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    put:
      tags:
        - recipes
      summary: Update a recipe.
      operationId: updateDish
      parameters:
        - name: id
          in: path
          description: recipe id.
          required: true
          style: simple
          explode: false
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeWithId'
        required: true
      responses:
        "200":
          description: recipe was updated.
          content:
            application-json:
              schema:
                $ref: '#/components/schemas/Recipe'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "403":
          description: Forbidden. Only admin can update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "404":
          description: recipe not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "500":
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - bearerAuth: []
    delete:
      tags:
        - recipes
      summary: Deletes a recipe
      description: '**Admin** endpoint'
      operationId: deleteRecipe
      parameters:
        - name: id
          in: path
          description: Recipe id to delete
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: OK
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "403":
          description: Forbidden. Only admin can update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "404":
          description: recipe not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "500":
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - bearerAuth: []
  /recipes/rand:
    get:
      tags:
        - recipes
      summary: Display 3 random recipes from Spooncaular API
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecipeWithId'
        "404":
          description: recipe not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "500":
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /recipes/search:
    get:
      tags:
        - recipes
      summary: Search for recipes from Spooncaular API
      parameters:
        - name: query
          in: query
          required: true
          style: form
          explode: true
          schema:
            type: string
            example: sushi
        - name: cuisine
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: string
            example: Japanese
        - name: diet
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: string
            example: vegetarian
        - name: intolerances
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: string
            example: onion
        - name: number
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: integer
            format: int64
            example: 5
        - name: instructionsRequired
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: boolean
            format: int64
            example: true
        - name: offset
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: integer
            format: int16
            example: 1
        - name: page
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: integer
            format: int16
            example: 1
        - name: limit
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: integer
            format: int16
            example: 2
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecipeWithId'
        "404":
          description: recipe not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "500":
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /recipe/{recipeId}/ingredients/{ingredientId}:
    put:
      tags:
        - recipes
      summary: Update the amount of ingredient in a recipe
      description: Update the amount of ingredient in a recipe
      operationId: updateIngredientAmount
      parameters:
        - name: recipeId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: ingredientId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/addIngredientToRecipeBody'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngredientInRecipe'
        "400":
          description: Bad request
        "404":
          description: Object not found
      security:
        - bearerAuth: []
    post:
      tags:
        - recipes
      summary: Add an ingredient to a recipe
      description: Add an amount of an ingredient to a given recipe
      operationId: addIngredientToRecipe
      parameters:
        - name: recipeId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: ingredientId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/addIngredientToRecipeBody'
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngredientInRecipe'
        "400":
          description: Bad request
        "404":
          description: Object not found
      security:
        - bearerAuth: []
    delete:
      tags:
        - recipes
      summary: Remove ingredient from a recipe.
      description: Remove ingredient from a recipe.
      operationId: deleteIngredientFromRecipe
      parameters:
        - name: recipeId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: ingredientId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: OK
        "404":
          description: Object not found
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    RecipeWithId:
      allOf:
        - $ref: '#/components/schemas/Recipe'
        - type: object
          properties:
            id:
              type: integer
              format: int64
...
    Recipe:
      required:
        - aggregate_likes
        - extended_ingredients
        - gluten_free
        - image
        - instructions
        - ready_in_minutes
        - serving
        - title
        - vegan
        - vegetarian
      type: object
      properties:
        title:
          type: string
          example: Spaghetti Carbonara
        ready_in_minutes:
          type: number
          example: 25
        aggregate_likes:
          type: number
          example: 1217
        serving:
          type: number
          example: 6
        vegetarian:
          type: boolean
          example: true
        vegan:
          type: boolean
          example: true
        gluten_free:
          type: boolean
          example: false
        image:
          type: string
          example: https://spoonacular.com/recipeImages/535835-556x370.jpg
        instructions:
          type: string
          example: "1. cook meat \n 2. prepare salad"
        extended_ingredients:
          type: array
          items:
            $ref: '#/components/schemas/IngredientInRecipe'
...
    IngredientInRecipe:
      required:
        - amount
        - id
        - unit
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: sugar
        amount:
          type: number
          format: float
          example: 2.5
        unit:
          type: string
          example: tablespoon
    ApiResponse:
      type: object
      properties:
        message:
          type: string
          description: Response description.
...
  requestBodies:
    addIngredientToRecipeBody:
      content:
        application/json:
          schema:
            required:
              - amount
            type: object
            properties:
              amount:
                type: number
                format: float
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT